Sort now takes ascending/descending arguement. getTotalPlayTimeOfArtist functional. getAlbumWithMostTracks also functional

+++ Programming2Coursework/Programming2Coursework/Album.h
@@ -61,7 +61,7 @@ inline std::vector<Track> Album::getTracks() const {
 	return tracks;
 }
 inline int Album::getNumSongs() const {
-	return tracks.size();
+	return (int)tracks.size();
 }
 
 //-------------------- General Methods ------------------------------
